support for mailserver test in a running graylog backend
